Make a few data types smaller and fix naming issue

+++ lights/lights_constants.h
@@ -5,12 +5,12 @@
 #define COLOR_ORDER GRB
 
 struct Lights_constants{
-	static const unsigned NUMBER_OF_LEDS = 192;
-	static const unsigned LED_STRIP_DATA_PIN = 3;
-	static const byte MAX_BRIGHTNESS = 255 * .5;
-	static constexpr double LEDS_PER_INCH = 1.3333333;	
+	static const uint8_t NUMBER_OF_LEDS = 192;
+	static const uint8_t LED_STRIP_DATA_PIN = 3;
+	static const uint8_t MAX_BRIGHTNESS = 255 * .5;
+	static constexpr float LEDS_PER_INCH = 1.3333333;	
 
-	static const unsigned FRAME_LENGTH = 100;//ms
+	static const uint8_t FRAME_LENGTH = 100;//ms
 };
 
 #endif
